requirements.txt
configparser==4.0.2
Django==2.0.2
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
mysql-connector-python==8.0.16
mysqlclient==1.4.4
protobuf==3.10.0
PyMySQL==0.9.3
python-gnupg==0.4.5
pytz==2019.3
six==1.12.0
